# perf: optimize JS communication with lazy getters