--- a/crates/rspack_macros/src/lib.rs
+++ b/crates/rspack_macros/src/lib.rs
@@ -2,6 +2,7 @@
 
 mod hook;
 mod merge;
+mod napi;
 mod plugin;
 mod runtime_module;
 mod source_map_config;
@@ -60,4 +61,12 @@ pub fn merge_from_derive(input: proc_macro::TokenStream) -> proc_macro::TokenStr
     syn::Result::Err(err) => err.to_compile_error(),
   }
   .into()
+}
+
+#[proc_macro_attribute]
+pub fn generate_getters(
+  args: proc_macro::TokenStream,
+  tokens: proc_macro::TokenStream,
+) -> proc_macro::TokenStream {
+  napi::generate_getters(args, tokens)
 }