<!--
  Thank you for submitting a pull request!

  We appreciate the time and effort you have invested in making these changes. Please ensure that you provide enough information to allow others to review your pull request.

  Upon submission, your pull request will be automatically assigned with reviewers.

  If you want to learn more about contributing to this project, please visit: https://github.com/web-infra-dev/rspack/blob/main/CONTRIBUTING.md.
-->

## Summary

<!-- Can you explain the reasoning behind implementing this change? What problem or issue does this pull request resolve? -->

<!-- It would be helpful if you could provide any relevant context, such as GitHub issues or related discussions. -->

This optimization uses getters to lazily create JavaScript values in Rust.

This enhances performance, as illustrated by `JsStatsModule`. In `JsStatsModule`, fields such as `identifier`, `name`, and `nameForCondition` are not always read. They are accessed only when `module.built`, `module.codeGenerated`, or `options.cachedModules` is true. Therefore, creating these JavaScript values in Rust when they are not used is unnecessary and inefficient.

Below are the results for an internal project when `stats.all = true`:

Before Optimization:

```
context.getInner(compilation).getChunks(): 38.889s
factory.create(`${type}.chunks`, chunks, context): 18.691s
```

After Optimization:

```
context.getInner(compilation).getChunks(): 1.829s
factory.create(`${type}.chunks`, chunks, context): 24.547s
```

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).